requirements.txt
gunicorn==23.0.0
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
environs[django]==11.0.0
psycopg2-binary==2.9.9
Faker==27.0.0
django-taggit==6.0.0
django-jalali-date==1.1.3
python-slugify==8.0.4
deep-translator==1.11.4
pillow==10.4.0
social-auth-app-django==5.4.2
django-admin-persian-fonts==0.2